Move subscription information to config files
After updating Jenkins, the credentials started being passed as a
sequence of `*`. To fix that was needed to just enter in the
configuration page for the provisioning jobs and click on save right
way, after that the credentials magically started working again.

To fix the above issue, the subscription was moved to a config file
which is sourced where needed (provisioning jobs as
satellite6-installer). This way no manual work will be needed when
updating the jobs using jenkins-job-builder. Also the credentials
information will be visible by jenkins admins which will make easier to
know what value they are set and update if needed.