Added support to reset sat6 instance using snapshot
a) This will use the target_image of the provisioning job as the
source_image for the Tier and rhai jobs.
b) Will re-provision a fresh sat6 instance each time a new Tier
job is triggered, by first destroying, undefining and deleting the
previous sat6 instance.
c) The idea is to use snapshot restore feature to do this.