Remove --boxed when calling py.test
When using --boxed pytest-xdist will run setUpClass for each test and not for
each process as was expected. Without the --boxed option pytest-xdist will run
setUpClass for each process which should decrease the time needed to run
automation.

Thank @rochacbruno for doing a good research on who the --boxed option works.

Close #324