Use new XUNIT Importer for Polarion Test Runs.
We can now use the new XUNIT Importer to populate Test Runs in
Polarion. Right now this PR uses code from a branch in Betelgeuse, but
once the changes from this branch are merged into master we should
update this PR to make sure that we use a published version of
Betelgeuse (instead of using a checkout).