Make polarion jobs to push results easily during re-run
Added support so that jobs can easily pull in artifacts
parse them and upload results to polarion.

Up until now, the tier jobs used to trigger the polarion
jobs with right set of values and there was no easy way
to re-run these jobs.

With this PR it would help us to easily re-run the jobs.

Signed-off-by: Kedar Bidarkar <kbidarka@redhat.com>