ignore npm install failure instead of using --ignore-scripts (#1305)
`npm install` in foreman fails in the postinstall script, if the ruby
environment is not prepared yet. the original idea was to run it with
--ignore-scripts, thus ignoring the postinstall script, and then
re-running it again after the ruby environment is ready.

however, --ignore-scripts ignores the scripts of ALL node modules, but a
second `npm install` only re-runs the scripts of Foreman, not of its
dependencies.

instead call `npm install || true` thus ignoring that `npm install`
bailed out, hoping that it bailed in the Foreman postinstall script,
which will be executed on the next `npm install` run.