archive the current state when releasing to QA
instead of archiving the *previous* snap, archive the currently released
one. this allows us to drop the previousSnapVersion parameter. it also
makes the new snap directly available under its number and opens up the
possibility to have an AK pointing to the exact snap directly after
release.

Note: before merging this, we'd have to manually archive the currently
released snaps, as otherwise those would be lost.