move release-pipeline and release-qa jobs to a template
the use of `{empty}` in the included paths is a silly trick to make jjb
not try to replace template variables INSIDE the file. the correct
solution would be to use `include-raw-escape` on these files, but you
cannot pass multiple `include-*` to `dsl` and I did not feel like
patching jjb just for that.