Update betelgeuse testrun and testplan uploads to set in progress status
We are including some manual test cases in these runs, and don't want to set their status to Done.

It looks like the default behavior here for betelgeuse is to set Done status on testruns, as there is nothing specified currently.
Within polarion, setting the test plan to Done does not set all included testruns to Done state, so I don't think its the use of `status=done` on the satellite6-polarion-test-plan.py script call that is setting testrun status too.